backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
feedparser>=6.0.11
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.1
python-dateutil>=2.9.0
pyyaml>=6.0.1
apscheduler>=3.10.4
anthropic>=0.25.0
google-generativeai>=0.5.4
pytest>=8.1.1
pytest-asyncio>=0.23.6
